Update main.m
This is my starter code for the project. The questions from the instruction sheet that I don't think I covered are bracketed by four percent symbols. Feel free to change this to make it better. My decision for parts 2 and 3 came from trial and error, and my decision for part 4 was just a quick choice so the coding could be simple.